feat: 수정사항 전체 반영 및 리팩토링(#10)
