+++ src/components/common/Header/index.js
@@ -1,20 +1,18 @@
-import React, { useCallback } from 'react';
-import * as S from './style';
-import {useHistory} from 'react-router-dom'; 
+import React from 'react'
+import * as S from './style'
+import { Link } from 'react-router-dom'
 
 const Header = () => {
-
-  const history = useHistory(); 
-  const onLoginClick = useCallback(()=>{
-    history.push('/login'); 
-  }, []); 
-
   return (
     <S.HeaderWrapper>
-      <S.HeaderContent>Rubminds</S.HeaderContent>
-      <S.HeaderContent onClick={onLoginClick}>로그인</S.HeaderContent>
+      <S.HeaderContent>
+        <Link to="/">Rubminds</Link>
+      </S.HeaderContent>
+      <S.HeaderContent>
+        <Link to="/login">로그인</Link>
+      </S.HeaderContent>
     </S.HeaderWrapper>
-  );
-};
+  )
+}
 
-export default Header;
+export default Header
